Rename package path to com.futurewei.alcor.vpcmanager
