+++ vpcmanager/src/main/java/com/futurewei/alcor/vpcmanager/VpcManagerApplication.java
@@ -1,4 +1,4 @@
-package com.futurewei.vpcmanager;
+package com.futurewei.alcor.vpcmanager;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
