+++ vpcmanager/src/main/java/com/futurewei/alcor/vpcmanager/config/ControllerConfig.java
@@ -13,7 +13,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
         See the License for the specific language governing permissions and
         limitations under the License.
 */
-package com.futurewei.vpcmanager.config;
+package com.futurewei.alcor.vpcmanager.config;
 
 public class ControllerConfig {
