+++ vpcmanager/src/main/java/com/futurewei/alcor/vpcmanager/dao/VpcRepository.java
@@ -14,7 +14,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
         limitations under the License.
 */
 
-package com.futurewei.vpcmanager.dao;
+package com.futurewei.alcor.vpcmanager.dao;
 
 import com.futurewei.common.repo.ICache;
 import com.futurewei.common.repo.ICacheRepository;
@@ -22,7 +22,7 @@ import com.futurewei.common.service.CacheFactory;
 import com.futurewei.common.exception.CacheException;
 import com.futurewei.common.logging.Logger;
 import com.futurewei.common.logging.LoggerFactory;
-import com.futurewei.vpcmanager.entity.VpcState;
+import com.futurewei.alcor.vpcmanager.entity.VpcState;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.stereotype.Repository;
