+++ vpcmanager/src/main/java/com/futurewei/alcor/vpcmanager/entity/VpcState.java
@@ -14,7 +14,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
         limitations under the License.
 */
 
-package com.futurewei.vpcmanager.entity;
+package com.futurewei.alcor.vpcmanager.entity;
 
 import com.futurewei.common.entity.CustomerResource;
 import lombok.Data;
