+++ vpcmanager/src/main/java/com/futurewei/alcor/vpcmanager/entity/VpcStateJson.java
@@ -13,9 +13,8 @@ Licensed under the Apache License, Version 2.0 (the "License");
         See the License for the specific language governing permissions and
         limitations under the License.
 */
-package com.futurewei.vpcmanager.entity;
+package com.futurewei.alcor.vpcmanager.entity;
 
-import com.futurewei.vpcmanager.entity.VpcState;
 import lombok.Data;
 
 @Data
